+++ test/serde_test.exs
@@ -15,74 +15,79 @@ defmodule Snex.SerdeTest do
 
   describe "Elixir nil" do
     test "is encoded as None", %{env: env} do
-      assert {:ok, true} = Snex.pyeval(env, %{"n" => nil}, returning: "n is None")
+      assert {:ok, true} = Snex.pyeval(env, "return n is None", %{"n" => nil})
     end
   end
 
   describe "Python None" do
     test "is decoded as nil", %{env: env} do
-      assert {:ok, nil} = Snex.pyeval(env, returning: "None")
+      assert {:ok, nil} = Snex.pyeval(env, "return None")
     end
   end
 
   describe "Elixir boolean" do
     test "true is encoded as True", %{env: env} do
-      assert {:ok, true} = Snex.pyeval(env, %{"b" => true}, returning: "b is True")
+      assert {:ok, true} = Snex.pyeval(env, "return b is True", %{"b" => true})
     end
 
     test "false is encoded as False", %{env: env} do
-      assert {:ok, true} = Snex.pyeval(env, %{"b" => false}, returning: "b is False")
+      assert {:ok, true} = Snex.pyeval(env, "return b is False", %{"b" => false})
     end
   end
 
   describe "Python boolean" do
     test "True is decoded as true", %{env: env} do
-      assert {:ok, true} = Snex.pyeval(env, returning: "True")
+      assert {:ok, true} = Snex.pyeval(env, "return True")
     end
 
     test "False is decoded as false", %{env: env} do
-      assert {:ok, false} = Snex.pyeval(env, returning: "False")
+      assert {:ok, false} = Snex.pyeval(env, "return False")
     end
   end
 
   describe "Elixir atom" do
     test "is encoded as snex.Atom", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"a" => :hello},
-                 returning: "type(a) is snex.Atom and a == 'hello'"
-               )
+               Snex.pyeval(env, "return type(a) is snex.Atom and a == 'hello'", %{"a" => :hello})
     end
 
     test "when empty", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"a" => :""}, returning: "type(a) is snex.Atom and a == ''")
+               Snex.pyeval(env, "return type(a) is snex.Atom and a == ''", %{"a" => :""})
     end
 
     test "with unicode characters", %{env: env} do
       atom = :"zażółć gęślą jaźń!"
 
       assert {:ok, true} =
-               Snex.pyeval(env, %{"a" => atom},
-                 returning: "type(a) is snex.Atom and a == 'zażółć gęślą jaźń!'"
+               Snex.pyeval(
+                 env,
+                 "return type(a) is snex.Atom and a == 'zażółć gęślą jaźń!'",
+                 %{"a" => atom}
                )
     end
 
     test "conflicts with string map keys", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"x" => %{"a" => "hello", a: :world}},
-                 returning: "len(x) == 1 and (x['a'] == 'hello' or x['a'] == 'world')"
+               Snex.pyeval(
+                 env,
+                 "return len(x) == 1 and (x['a'] == 'hello' or x['a'] == 'world')",
+                 %{"x" => %{"a" => "hello", a: :world}}
                )
     end
 
     test "does not conflict with string map keys when `:distinct_atom` is used", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"x" => %{"a" => "hello", a: :world}},
-                 encoding_opts: [atom_as: :distinct_atom],
-                 returning: """
-                 len(x) == 2 \
-                 and x['a'] == 'hello' \
-                 and x[snex.DistinctAtom('a')] == snex.DistinctAtom('world')\
+               Snex.pyeval(
+                 env,
                  """
+                 return \
+                   len(x) == 2 \
+                   and x['a'] == 'hello' \
+                   and x[snex.DistinctAtom('a')] == snex.DistinctAtom('world')\
+                 """,
+                 %{"x" => %{"a" => "hello", a: :world}},
+                 encoding_opts: [atom_as: :distinct_atom]
                )
     end
 
@@ -91,8 +96,10 @@ defmodule Snex.SerdeTest do
       assert byte_size(atom_str) > 255
 
       assert {:ok, true} =
-               Snex.pyeval(env, %{"a" => String.to_atom(atom_str)},
-                 returning: "type(a) is snex.Atom and len(a) == 255"
+               Snex.pyeval(
+                 env,
+                 "return type(a) is snex.Atom and len(a) == 255",
+                 %{"a" => String.to_atom(atom_str)}
                )
     end
 
@@ -114,17 +121,22 @@ defmodule Snex.SerdeTest do
 
       # interpreter default
       assert {:ok, true} =
-               Snex.pyeval(env, %{"a" => atom},
-                 returning: """
+               Snex.pyeval(
+                 env,
+                 """
+                 return \
                  type(a) is snex.DistinctAtom \
                  and a == snex.DistinctAtom('hello') \
                  and a != 'hello'\
-                 """
+                 """,
+                 %{"a" => atom}
                )
 
       assert {:ok, true} =
-               Snex.pyeval(env,
-                 returning: """
+               Snex.pyeval(
+                 env,
+                 """
+                 return \
                  type(var_default) is snex.DistinctAtom \
                  and var_default == snex.DistinctAtom('hello') \
                  and var_default != 'hello'\
@@ -136,15 +148,18 @@ defmodule Snex.SerdeTest do
             distinct_atom: {"snex.DistinctAtom", "snex.DistinctAtom('hello')"}
           ] do
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"a" => atom},
-                   encoding_opts: [atom_as: as],
-                   returning: "type(a) is #{type} and a == #{expected}"
+                 Snex.pyeval(
+                   env,
+                   "return type(a) is #{type} and a == #{expected}",
+                   %{"a" => atom},
+                   encoding_opts: [atom_as: as]
                  )
 
         assert {:ok, true} =
-                 Snex.pyeval(env,
-                   encoding_opts: [atom_as: as],
-                   returning: "type(var_#{as}) is #{type} and var_#{as} == #{expected}"
+                 Snex.pyeval(
+                   env,
+                   "return type(var_#{as}) is #{type} and var_#{as} == #{expected}",
+                   encoding_opts: [atom_as: as]
                  )
       end
     end
@@ -152,21 +167,21 @@ defmodule Snex.SerdeTest do
 
   describe "Python snex.Atom" do
     test "is decoded as atom", %{env: env} do
-      assert {:ok, :hello} = Snex.pyeval(env, returning: "snex.Atom('hello')")
+      assert {:ok, :hello} = Snex.pyeval(env, "return snex.Atom('hello')")
     end
 
     test "when empty", %{env: env} do
-      assert {:ok, :""} = Snex.pyeval(env, returning: "snex.Atom('')")
+      assert {:ok, :""} = Snex.pyeval(env, "return snex.Atom('')")
     end
 
     test "can be 255 characters long", %{env: env} do
       expected = String.duplicate("a", 255) |> String.to_atom()
-      assert {:ok, ^expected} = Snex.pyeval(env, returning: "snex.Atom('a' * 255)")
+      assert {:ok, ^expected} = Snex.pyeval(env, "return snex.Atom('a' * 255)")
     end
 
     test "cannot be longer than 255 characters", %{env: env} do
       assert {:error, %Snex.Error{code: :python_runtime_error, reason: "atom too long: " <> _}} =
-               Snex.pyeval(env, returning: "snex.Atom('a' * 256)")
+               Snex.pyeval(env, "return snex.Atom('a' * 256)")
     end
 
     test "is counted by character length", %{env: env} do
@@ -175,10 +190,10 @@ defmodule Snex.SerdeTest do
       expected_atom = String.to_atom(expected_string)
 
       assert byte_size(expected_string) > 255
-      assert {:ok, ^expected_atom} = Snex.pyeval(env, returning: "snex.Atom('ą' * 255)")
+      assert {:ok, ^expected_atom} = Snex.pyeval(env, "return snex.Atom('ą' * 255)")
 
       assert {:error, %Snex.Error{code: :python_runtime_error, reason: "atom too long: " <> _}} =
-               Snex.pyeval(env, returning: "snex.Atom('ą' * 256)")
+               Snex.pyeval(env, "return snex.Atom('ą' * 256)")
     end
   end
 
@@ -207,7 +222,7 @@ defmodule Snex.SerdeTest do
             -(1 <<< (8 * 0xFF + 1))
           ] do
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"i" => int}, returning: "type(i) is int and i == #{int}")
+                 Snex.pyeval(env, "return type(i) is int and i == #{int}", %{"i" => int})
       end
     end
   end
@@ -228,7 +243,7 @@ defmodule Snex.SerdeTest do
             1 <<< (8 * 0xFF),
             -(1 <<< (8 * 0xFF))
           ] do
-        assert {:ok, ^int} = Snex.pyeval(env, returning: "#{int}")
+        assert {:ok, ^int} = Snex.pyeval(env, "return #{int}")
       end
     end
   end
@@ -237,9 +252,7 @@ defmodule Snex.SerdeTest do
     test "is encoded as Python float", %{env: env} do
       for float <- [-1.25, 1.25, 0.0, -0.0, Float.min_finite(), Float.max_finite()] do
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"f" => float},
-                   returning: "type(f) is float and f == #{float}"
-                 )
+                 Snex.pyeval(env, "return type(f) is float and f == #{float}", %{"f" => float})
       end
     end
   end
@@ -248,30 +261,38 @@ defmodule Snex.SerdeTest do
     test "when number", %{env: env} do
       for float <- [-1.25, 1.25, 0.0, -0.0, Float.min_finite(), Float.max_finite()] do
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"f" => Snex.Serde.float(float)},
-                   returning: "type(f) is float and f == #{float}"
+                 Snex.pyeval(
+                   env,
+                   "return type(f) is float and f == #{float}",
+                   %{"f" => Snex.Serde.float(float)}
                  )
       end
     end
 
     test "when infinity", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"f" => Snex.Serde.float(:inf)},
-                 returning: "type(f) is float and math.isinf(f) and f > 0"
+               Snex.pyeval(
+                 env,
+                 "return type(f) is float and math.isinf(f) and f > 0",
+                 %{"f" => Snex.Serde.float(:inf)}
                )
     end
 
     test "when negative infinity", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"f" => Snex.Serde.float(:"-inf")},
-                 returning: "type(f) is float and math.isinf(f) and f < 0"
+               Snex.pyeval(
+                 env,
+                 "return type(f) is float and math.isinf(f) and f < 0",
+                 %{"f" => Snex.Serde.float(:"-inf")}
                )
     end
 
     test "when NaN", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"f" => Snex.Serde.float(:nan)},
-                 returning: "type(f) is float and math.isnan(f)"
+               Snex.pyeval(
+                 env,
+                 "return type(f) is float and math.isnan(f)",
+                 %{"f" => Snex.Serde.float(:nan)}
                )
     end
   end
@@ -279,20 +300,20 @@ defmodule Snex.SerdeTest do
   describe "Python float" do
     test "is decoded as Elixir float", %{env: env} do
       for float <- [1.25, 0.0, -0.0, Float.min_finite(), Float.max_finite()] do
-        assert {:ok, ^float} = Snex.pyeval(env, returning: "#{float}")
+        assert {:ok, ^float} = Snex.pyeval(env, "return #{float}")
       end
     end
 
     test "is decoded as :inf when infinity", %{env: env} do
-      assert {:ok, :inf} = Snex.pyeval(env, returning: "float('inf')")
+      assert {:ok, :inf} = Snex.pyeval(env, "return float('inf')")
     end
 
     test ~s'is decoded as :"-inf" when negative infinity', %{env: env} do
-      assert {:ok, :"-inf"} = Snex.pyeval(env, returning: "float('-inf')")
+      assert {:ok, :"-inf"} = Snex.pyeval(env, "return float('-inf')")
     end
 
     test "is decoded as :nan when NaN", %{env: env} do
-      assert {:ok, :nan} = Snex.pyeval(env, returning: "float('NaN')")
+      assert {:ok, :nan} = Snex.pyeval(env, "return float('NaN')")
     end
   end
 
@@ -302,18 +323,18 @@ defmodule Snex.SerdeTest do
         str = String.duplicate("a", strlen)
 
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"s" => str}, returning: "type(s) is str and s == '#{str}'")
+                 Snex.pyeval(env, "return type(s) is str and s == '#{str}'", %{"s" => str})
       end
     end
 
     test "round-trips unicode by default", %{env: env} do
       s = "zażółć gęślą jaźń"
-      assert {:ok, ^s} = Snex.pyeval(env, %{"s" => s}, returning: "s")
+      assert {:ok, ^s} = Snex.pyeval(env, "return s", %{"s" => s})
     end
 
     test "breaks when non-UTF8 by default", %{env: env} do
       assert {:error, %Snex.Error{code: :python_runtime_error}} =
-               Snex.pyeval(env, %{"s" => <<0xFF>>}, returning: "s")
+               Snex.pyeval(env, "return s", %{"s" => <<0xFF>>})
     end
 
     test "is encoded according to `encoding_opts`" do
@@ -332,12 +353,10 @@ defmodule Snex.SerdeTest do
 
       # interpreter default
       assert {:ok, true} =
-               Snex.pyeval(env, %{"b" => bin}, returning: "type(b) is bytes and b == b'hello'")
+               Snex.pyeval(env, "return type(b) is bytes and b == b'hello'", %{"b" => bin})
 
       assert {:ok, true} =
-               Snex.pyeval(env,
-                 returning: "type(var_default) is bytes and var_default == b'hello'"
-               )
+               Snex.pyeval(env, "return type(var_default) is bytes and var_default == b'hello'")
 
       for {type, expected} <- [
             bytes: "b'hello'",
@@ -345,15 +364,18 @@ defmodule Snex.SerdeTest do
             str: "'hello'"
           ] do
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"b" => bin},
-                   encoding_opts: [binary_as: type],
-                   returning: "type(b) is #{type} and b == #{expected}"
+                 Snex.pyeval(
+                   env,
+                   "return type(b) is #{type} and b == #{expected}",
+                   %{"b" => bin},
+                   encoding_opts: [binary_as: type]
                  )
 
         assert {:ok, true} =
-                 Snex.pyeval(env,
-                   encoding_opts: [binary_as: type],
-                   returning: "type(var_#{type}) is #{type} and var_#{type} == #{expected}"
+                 Snex.pyeval(
+                   env,
+                   "return type(var_#{type}) is #{type} and var_#{type} == #{expected}",
+                   encoding_opts: [binary_as: type]
                  )
       end
     end
@@ -363,12 +385,12 @@ defmodule Snex.SerdeTest do
     test "is decoded as Elixir binary", %{env: env} do
       for strlen <- [0, 1, 255, 256] do
         str = String.duplicate("a", strlen)
-        assert {:ok, ^str} = Snex.pyeval(env, returning: "'#{str}'")
+        assert {:ok, ^str} = Snex.pyeval(env, "return '#{str}'")
       end
     end
 
     test "when unicode", %{env: env} do
-      assert {:ok, "zażółć gęślą jaźń"} = Snex.pyeval(env, returning: "'zażółć gęślą jaźń'")
+      assert {:ok, "zażółć gęślą jaźń"} = Snex.pyeval(env, "return 'zażółć gęślą jaźń'")
     end
   end
 
@@ -378,16 +400,20 @@ defmodule Snex.SerdeTest do
         bin = :binary.copy(<<171>>, len)
 
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"b" => Snex.Serde.binary(bin)},
-                   returning: "type(b) is bytes and b == bytes([171]) * #{len}"
+                 Snex.pyeval(
+                   env,
+                   "return type(b) is bytes and b == bytes([171]) * #{len}",
+                   %{"b" => Snex.Serde.binary(bin)}
                  )
       end
     end
 
     test "wraps iodata", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"b" => Snex.Serde.binary([<<1, 2, 3>>, <<4>>])},
-                 returning: "type(b) is bytes and b == bytes([1, 2, 3, 4])"
+               Snex.pyeval(
+                 env,
+                 "return type(b) is bytes and b == bytes([1, 2, 3, 4])",
+                 %{"b" => Snex.Serde.binary([<<1, 2, 3>>, <<4>>])}
                )
     end
 
@@ -400,8 +426,10 @@ defmodule Snex.SerdeTest do
             bytes: "b'hello'"
           ] do
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"b" => Snex.Serde.binary(bin, type)},
-                   returning: "type(b) is #{type} and b == #{expected}"
+                 Snex.pyeval(
+                   env,
+                   "return type(b) is #{type} and b == #{expected}",
+                   %{"b" => Snex.Serde.binary(bin, type)}
                  )
       end
     end
@@ -413,9 +441,10 @@ defmodule Snex.SerdeTest do
         bin = :binary.copy(<<171>>, len)
 
         assert {:ok, {^bin, ^bin, ^bin}} =
-                 Snex.pyeval(env, "b = bytearray([171] * #{len})",
-                   returning: "(b, bytes(b), memoryview(b))"
-                 )
+                 Snex.pyeval(env, """
+                 b = bytearray([171] * #{len})
+                 return (b, bytes(b), memoryview(b))
+                 """)
       end
     end
   end
@@ -427,8 +456,10 @@ defmodule Snex.SerdeTest do
       term = {:ok, pid, ref, env, %{a: 1, b: [2, 3]}}
 
       assert {:ok, {true, ^term}} =
-               Snex.pyeval(env, %{"t" => Snex.Serde.term(term)},
-                 returning: "(type(t) is snex.Term, t)"
+               Snex.pyeval(
+                 env,
+                 "return (type(t) is snex.Term, t)",
+                 %{"t" => Snex.Serde.term(term)}
                )
     end
   end
@@ -436,35 +467,31 @@ defmodule Snex.SerdeTest do
   describe "Elixir list" do
     test "is encoded as Python list", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"l" => [1, 2, 3]},
-                 returning: "type(l) is list and l == [1, 2, 3]"
-               )
+               Snex.pyeval(env, "return type(l) is list and l == [1, 2, 3]", %{"l" => [1, 2, 3]})
     end
 
     test "when empty", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"l" => []}, returning: "type(l) is list and l == []")
+               Snex.pyeval(env, "return type(l) is list and l == []", %{"l" => []})
     end
 
     test "when nested", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"l" => [[], [[]]]},
-                 returning: "type(l) is list and l == [[], [[]]]"
-               )
+               Snex.pyeval(env, "return type(l) is list and l == [[], [[]]]", %{"l" => [[], [[]]]})
     end
   end
 
   describe "Python list" do
     test "is decoded as Elixir list", %{env: env} do
-      assert {:ok, [1, 2, 3]} = Snex.pyeval(env, returning: "[1, 2, 3]")
+      assert {:ok, [1, 2, 3]} = Snex.pyeval(env, "return [1, 2, 3]")
     end
 
     test "when empty", %{env: env} do
-      assert {:ok, []} = Snex.pyeval(env, returning: "[]")
+      assert {:ok, []} = Snex.pyeval(env, "return []")
     end
 
     test "when nested", %{env: env} do
-      assert {:ok, [[], [[]]]} = Snex.pyeval(env, returning: "[[], [[]]]")
+      assert {:ok, [[], [[]]]} = Snex.pyeval(env, "return [[], [[]]]")
     end
   end
 
@@ -475,9 +502,10 @@ defmodule Snex.SerdeTest do
         tuple = List.to_tuple(list)
 
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"t" => tuple},
-                   returning:
-                     "type(t) is tuple and t == tuple(#{inspect(list, limit: :infinity)})"
+                 Snex.pyeval(
+                   env,
+                   "return type(t) is tuple and t == tuple(#{inspect(list, limit: :infinity)})",
+                   %{"t" => tuple}
                  )
       end
     end
@@ -490,7 +518,7 @@ defmodule Snex.SerdeTest do
         tuple = List.to_tuple(list)
 
         assert {:ok, ^tuple} =
-                 Snex.pyeval(env, returning: "tuple(#{inspect(list, limit: :infinity)})")
+                 Snex.pyeval(env, "return tuple(#{inspect(list, limit: :infinity)})")
       end
     end
   end
@@ -500,45 +528,50 @@ defmodule Snex.SerdeTest do
       m = %{"x" => 1, "nested" => %{z: 3}, y: 2}
 
       assert {:ok, true} =
-               Snex.pyeval(env, %{"m" => m},
-                 returning:
-                   "type(m) is dict and m['x'] == 1 and m['y'] == 2 and m['nested']['z'] == 3"
+               Snex.pyeval(
+                 env,
+                 "return type(m) is dict and m['x'] == 1 and m['y'] == 2 and m['nested']['z'] == 3",
+                 %{"m" => m}
                )
     end
 
     test "when empty", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"m" => %{}}, returning: "type(m) is dict and m == {}")
+               Snex.pyeval(env, "return type(m) is dict and m == {}", %{"m" => %{}})
     end
   end
 
   describe "Python dict" do
     test "is decoded as Elixir map", %{env: env} do
-      assert {:ok, %{"x" => 1, "y" => 2}} = Snex.pyeval(env, returning: "{'x': 1, 'y': 2}")
+      assert {:ok, %{"x" => 1, "y" => 2}} = Snex.pyeval(env, "return {'x': 1, 'y': 2}")
     end
 
     test "when empty", %{env: env} do
-      assert {:ok, %{}} = Snex.pyeval(env, returning: "{}")
+      assert {:ok, %{}} = Snex.pyeval(env, "return {}")
     end
 
     test "with snex.Atom keys", %{env: env} do
       assert {:ok, %{x: 1, y: 2}} =
-               Snex.pyeval(env, returning: "{snex.Atom('x'): 1, snex.Atom('y'): 2}")
+               Snex.pyeval(env, "return {snex.Atom('x'): 1, snex.Atom('y'): 2}")
     end
   end
 
   describe "Elixir MapSet" do
     test "is encoded as Python set", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"s" => MapSet.new([1, 2, 3])},
-                 returning: "type(s) is set and s == {1, 2, 3}"
+               Snex.pyeval(
+                 env,
+                 "return type(s) is set and s == {1, 2, 3}",
+                 %{"s" => MapSet.new([1, 2, 3])}
                )
     end
 
     test "when empty", %{env: env} do
       assert {:ok, true} =
-               Snex.pyeval(env, %{"s" => MapSet.new()},
-                 returning: "type(s) is set and s == set()"
+               Snex.pyeval(
+                 env,
+                 "return type(s) is set and s == set()",
+                 %{"s" => MapSet.new()}
                )
     end
 
@@ -560,26 +593,32 @@ defmodule Snex.SerdeTest do
 
       # interpreter default
       assert {:ok, true} =
-               Snex.pyeval(env, %{"s" => set},
-                 returning: "type(s) is frozenset and s == {1, 2, 3}"
+               Snex.pyeval(
+                 env,
+                 "return type(s) is frozenset and s == {1, 2, 3}",
+                 %{"s" => set}
                )
 
       assert {:ok, true} =
-               Snex.pyeval(env,
-                 returning: "type(var_default) is frozenset and var_default == {1, 2, 3}"
+               Snex.pyeval(
+                 env,
+                 "return type(var_default) is frozenset and var_default == {1, 2, 3}"
                )
 
       for type <- [:set, :frozenset] do
         assert {:ok, true} =
-                 Snex.pyeval(env, %{"s" => set},
-                   encoding_opts: [set_as: type],
-                   returning: "type(s) is #{type} and s == {1, 2, 3}"
+                 Snex.pyeval(
+                   env,
+                   "return type(s) is #{type} and s == {1, 2, 3}",
+                   %{"s" => set},
+                   encoding_opts: [set_as: type]
                  )
 
         assert {:ok, true} =
-                 Snex.pyeval(env,
-                   encoding_opts: [set_as: type],
-                   returning: "type(var_#{type}) is #{type} and var_#{type} == {1, 2, 3}"
+                 Snex.pyeval(
+                   env,
+                   "return type(var_#{type}) is #{type} and var_#{type} == {1, 2, 3}",
+                   encoding_opts: [set_as: type]
                  )
       end
     end
@@ -590,12 +629,12 @@ defmodule Snex.SerdeTest do
       expected_set = MapSet.new([1, 2, 3])
 
       assert {:ok, {^expected_set, ^expected_set}} =
-               Snex.pyeval(env, returning: "(set([1, 2, 3]), frozenset([1, 2, 3]))")
+               Snex.pyeval(env, "return (set([1, 2, 3]), frozenset([1, 2, 3]))")
     end
 
     test "empty set decodes as empty MapSet", %{env: env} do
       empty_set = MapSet.new()
-      assert {:ok, {^empty_set, ^empty_set}} = Snex.pyeval(env, returning: "(set(), frozenset())")
+      assert {:ok, {^empty_set, ^empty_set}} = Snex.pyeval(env, "return (set(), frozenset())")
     end
   end
 
@@ -604,20 +643,23 @@ defmodule Snex.SerdeTest do
       uri = URI.parse("https://example.com/path?q=1")
 
       assert {:ok, true} =
-               Snex.pyeval(env, %{"uri" => uri},
-                 returning: """
+               Snex.pyeval(
+                 env,
+                 """
+                 return \
                  type(uri) is dict \
                  and all(type(k) is snex.Atom for k in uri.keys()) \
                  and type(uri['__struct__']) is snex.Atom \
                  and uri['__struct__'] == 'Elixir.URI' \
                  and uri[snex.Atom('scheme')] == 'https'\
-                 """
+                 """,
+                 %{"uri" => uri}
                )
     end
 
     test "round-trips (generic struct encoding)", %{env: env} do
       uri = URI.parse("https://example.com/path?q=1")
-      assert {:ok, ^uri} = Snex.pyeval(env, %{"uri" => uri}, returning: "uri")
+      assert {:ok, ^uri} = Snex.pyeval(env, "return uri", %{"uri" => uri})
     end
 
     test "with custom Snex.Serde.Encoder", %{env: env} do
@@ -625,8 +667,10 @@ defmodule Snex.SerdeTest do
       date = ~D[2025-12-28]
 
       assert {:ok, true} =
-               Snex.pyeval(env, %{"d" => date},
-                 returning: "type(d) is datetime.date and d == datetime.date(2025, 12, 28)"
+               Snex.pyeval(
+                 env,
+                 "return type(d) is datetime.date and d == datetime.date(2025, 12, 28)",
+                 %{"d" => date}
                )
     end
   end
@@ -649,7 +693,7 @@ defmodule Snex.SerdeTest do
                snex.set_custom_encoder(custom_encoder)
                """)
 
-      assert {:ok, ~D[2025-12-28]} = Snex.pyeval(env, returning: "datetime.date(2025, 12, 28)")
+      assert {:ok, ~D[2025-12-28]} = Snex.pyeval(env, "return datetime.date(2025, 12, 28)")
     end
   end
 
@@ -663,7 +707,7 @@ defmodule Snex.SerdeTest do
         "bool" => true
       }
 
-      assert {:ok, ^nested} = Snex.pyeval(env, %{"n" => nested}, returning: "n")
+      assert {:ok, ^nested} = Snex.pyeval(env, "return n", %{"n" => nested})
     end
   end
 end